功能优化

+++ src/main/java/com/qingshixun/service/impl/DepartmentService.java
@@ -6,8 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.qingshixun.core.ResponseData;
 import com.qingshixun.dao.IDepartmentDao;
 import com.qingshixun.service.IDepartmentService;
+import com.qingshixun.service.IUserService;
 import com.qingshixun.model.Department;
 
 @Service("departmentService")
@@ -16,6 +18,9 @@ public class DepartmentService implements IDepartmentService {
 
 	@Autowired
 	private IDepartmentDao departmentDao;
+	
+	@Autowired
+	private IUserService userService;
 
 	/**
 	 * 查询部门信息
@@ -38,8 +43,15 @@ public class DepartmentService implements IDepartmentService {
 	 * 通过参数departmentId查询并删除对应部门信息
 	 */
 	@Override
-	public void deleteDepartment(int departmentId) {
+	public ResponseData deleteDepartment(int departmentId) {
+		ResponseData responseData = new ResponseData();
+		if (!userService.removeUserDepartment(departmentId)) {
+			responseData.setError("部门已被使用，无法删除！");
+			return responseData;
+		}
 		departmentDao.deleteDepartment(departmentId);
+		responseData.setError("删除成功");
+		return responseData;
 	}
 
 	/**
